seup decorator without msw
