+++ .storybook/preview.tsx
@@ -1,0 +1,33 @@
+import type { Preview } from '@storybook/react'
+
+import { initialize, mswDecorator } from 'msw-storybook-addon'
+
+import { Provider as ReduxStoreProvider } from 'react-redux'
+import { store } from '../src/redux/store'
+import { handlers } from '../mocks/handlers'
+import '../src/index.css'
+
+// Initialize MSW
+initialize()
+
+const preview: Preview = {
+  parameters: {
+    actions: { argTypesRegex: '^on[A-Z].*' },
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/,
+      },
+    },
+  },
+  decorators: [
+    mswDecorator,
+    (Story) => (
+      <ReduxStoreProvider store={store}>
+        <Story />
+      </ReduxStoreProvider>
+    ),
+  ],
+}
+
+export default preview
