+++ tsconfig.json
@@ -18,6 +18,6 @@
     "strict": true,
     "target": "ES2020"
   },
-  "include": ["src", "validator", "@types", "mocks", "lib"],
-  "exclude": ["server, build", "server_build", ".storybook", "storybook-static"]
+  "include": ["src", "validator", "@types", "mocks", "lib", ".storybook"],
+  "exclude": ["server, build", "server_build", "storybook-static"]
 }
